// src/app/api/swagger/models/api-application-state-enum-2.ts
/* tslint:disable */
/* eslint-disable */
export enum ApiApplicationStateEnum2 {
  Formation = 1,
  Complete = 2,
  Rejected = 3,
  Approved = 4,
  Accepted = 5,
  RejectedByPart1 = 6,
  AcceptedByPart1 = 7,
  NotImplemented = 8,
  NotConsideredPart2 = 9,
  RejectedByPart2 = 10,
  AcceptedByPart2 = 11,
  Loser = 12,
  Winner = 13,
  ContractTerminated = 14,
  Withdrawn = 15,
  NotActual = 16,
  WaitingForFinalProposal = 17,
  FinalProposalSubmitted = 18
}